s/usize/u16/

+++ wasm/src/cpu.rs
@@ -19,7 +19,7 @@ impl CPU {
         self.execute(byte);
     }
 
-    fn execute(&mut self, byte: u8) -> usize {
+    fn execute(&mut self, byte: u8) -> u16 {
         let bits = (
             byte & 0b00000001,
             (byte & 0b00000010) >> 1,
@@ -43,27 +43,27 @@ impl CPU {
         pc
     }
 
-    fn op_add(&mut self, byte: u8) -> usize {
+    fn op_add(&mut self, byte: u8) -> u16 {
         self.registers.a += self.register_8_get(byte);
         self.registers.pc + 1
     }
 
-    fn op_ld_a_rp(&mut self, byte: u8) -> usize {
-        self.registers.a = self.bus.get_byte(self.register_16_get(byte) as usize);
+    fn op_ld_a_rp(&mut self, byte: u8) -> u16 {
+        self.registers.a = self.bus.get_byte(self.register_16_get(byte));
         self.registers.pc + 1
     }
 
-    fn op_ld_r_n(&mut self, byte: u8) -> usize {
+    fn op_ld_r_n(&mut self, byte: u8) -> u16 {
         self.register_8_set(byte, self.bus.get_byte(self.registers.pc + 1));
         self.registers.pc + 2
     }
 
-    fn op_ld_r_r(&mut self, byte: u8) -> usize {
+    fn op_ld_r_r(&mut self, byte: u8) -> u16 {
         self.register_8_set(byte, self.register_8_get(byte));
         self.registers.pc + 1
     }
 
-    fn op_ld_rp_a(&mut self, byte: u8) -> usize {
+    fn op_ld_rp_a(&mut self, byte: u8) -> u16 {
         self.register_16_set(byte, self.registers.a);
         self.registers.pc + 1
     }
@@ -80,9 +80,9 @@ impl CPU {
 
     fn register_16_set(&mut self, byte: u8, value: u8) {
         match byte & 0b110000 {
-            0b000000 => self.bus.set_byte(self.registers.get_bc() as usize, value),
-            0b010000 => self.bus.set_byte(self.registers.get_de() as usize, value),
-            0b100000 => self.bus.set_byte(self.registers.get_hl() as usize, value),
+            0b000000 => self.bus.set_byte(self.registers.get_bc(), value),
+            0b010000 => self.bus.set_byte(self.registers.get_de(), value),
+            0b100000 => self.bus.set_byte(self.registers.get_hl(), value),
             0b110000 => self.bus.set_byte(self.registers.sp, value),
             _ => panic!("never reach"),
         }
@@ -96,7 +96,7 @@ impl CPU {
             0b011 => self.registers.e,
             0b100 => self.registers.h,
             0b101 => self.registers.l,
-            0b110 => (self.bus.get_byte(self.registers.get_hl() as usize)),
+            0b110 => (self.bus.get_byte(self.registers.get_hl())),
             0b111 => self.registers.a,
             _ => panic!("never reach"),
         }
@@ -110,7 +110,7 @@ impl CPU {
             0b011000 => self.registers.e = value,
             0b100000 => self.registers.h = value,
             0b101000 => self.registers.l = value,
-            0b110000 => self.bus.set_byte(self.registers.get_hl() as usize, value),
+            0b110000 => self.bus.set_byte(self.registers.get_hl(), value),
             0b111000 => self.registers.a = value,
             _ => panic!("invalid destination register"),
         }
